restructured submodule types
