+++ src/types/mod.rs
@@ -1,0 +1,9 @@
+//! Types used in the Matrix API.
+//!
+//! Will be better documented in the future; for now,
+//! refer to the official API docs for info on what fields mean.
+
+pub mod messages;
+pub mod content;
+pub mod events;
+pub mod replies;
