+++ src/types/replies.rs
@@ -1,0 +1,55 @@
+use std::collections::HashMap;
+use ::types::events::{Events};
+
+/// Information about a room's events.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Room {
+    #[serde(default)]
+    pub state: Events,
+    #[serde(default)]
+    pub timeline: Events,
+}
+/// The `rooms` component of a `SyncReply`.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct SyncRooms {
+    #[serde(default)]
+    pub join: HashMap<String, Room>,
+    #[serde(default)]
+    pub invite: HashMap<String, Room>,
+    #[serde(default)]
+    pub leave: HashMap<String, Room>
+}
+/// The reply obtained from `sync()`.
+#[derive(Deserialize, Debug)]
+pub struct SyncReply {
+    pub next_batch: String,
+    pub rooms: SyncRooms
+}
+/// The reply obtained from `/send`.
+#[derive(Deserialize, Debug)]
+pub struct SendReply {
+    pub event_id: String
+}
+/// The reply obtained from `upload()`.
+#[derive(Deserialize, Debug)]
+pub struct UploadReply {
+    pub content_uri: String
+}
+/// The reply obtained from `/join`.
+#[derive(Deserialize, Debug)]
+pub struct JoinReply {
+    pub room_id: String
+}
+/// The reply obtained from `/login`.
+#[derive(Deserialize, Debug)]
+pub struct LoginReply {
+    pub user_id: String,
+    pub access_token: String,
+    pub home_server: String
+}
+/// The reply obtained when something's gone wrong.
+#[derive(Deserialize, Debug)]
+pub struct BadRequestReply {
+    pub errcode: String,
+    pub error: String
+}
